+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Possible log types:
 ### Unreleased
 
 - [added] Added the `HumitidySensor` and `PowerConsumptionSensor` sensor types.
+- [added] Added the `timezone` key under `location` section.
 
 ### v0.8.1 (2021-07-06)